Update/growth ui

See merge request wallet/proton-wallet-app!281

+++ lib/rust/proton_api/price_graph.dart
@@ -8,7 +8,7 @@ import 'package:flutter_rust_bridge/flutter_rust_bridge_for_generated.dart';
 import 'user_settings.dart';
 
 class DataPoint {
-  final int exchangeRate;
+  final BigInt exchangeRate;
   final int? cents;
   final BigInt timestamp;
